RavenDB-16926 - Move WaitForRaftCommandToBeAppliedInClusterAsync to ClusterTestBase2
